requirements.txt
IPython==7.18.0
hashids==1.3.1
alembic==1.4.2
antlr4-python3-runtime==4.8
babel==2.8.0
cython==0.29.21
dask==2.24.0
distributed==2.24.0
cloudpickle==1.5.0
gcloud==0.18.3
h5py==2.10.0
locket==0.2.0
Mako==1.1.3
Markdown==3.2.2
mock==4.0.2
msgpack==1.0.0
numpy==1.19.1
pandas==1.1.1
parsedatetime==2.6
pep8==1.7.1
pyarrow==1.0.1
requests==2.24.0
scipy==1.5.2
sortedcollections==1.2.1
sortedcontainers==2.2.2
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
tox==3.19.0
flask==1.1.2
flask-cors==3.0.8
dash==1.15.0
dash-bootstrap-components==0.10.5
humanize
pytz
yaml
